// packages/contracts-ts/src/generated/rino-registry-v1.schema.ts
export const rinoRegistryV1Schema = {
// ...
  "$defs": {
// ...
    "PrimitiveTypeV1": {
      "title": "PrimitiveTypeV1",
      "type": "object",
      "additionalProperties": false,
      "required": [
        "kind"
      ],
      "properties": {
        "kind": {
          "$ref": "#/$defs/PrimitiveTypeKindV1"
        }
      }
    },
// ...
    "WorkflowTemplateV1": {
      "title": "WorkflowTemplateV1",
      "type": "object",
      "additionalProperties": false,
      "required": [
        "templateKey",
        "titleKey",
        "descriptionKey",
        "iconKey",
        "nodes"
      ],
      "properties": {
// ...
        "edges": {
          "type": "array",
          "items": {
            "$ref": "#/$defs/TemplateEdgeV1"
          },
          "maxItems": 128
        },
// ...
      }
    }
  }
} as const;